+++ djem/__init__.py
@@ -1,6 +1,7 @@
 __version__ = '0.5.0a2'
 
 # Shortcuts to goodies
-from djem.misc import AjaxResponse, get_page  # NOQA
+from djem.ajax import AjaxResponse  # NOQA
+from djem.pagination import get_page  # NOQA
 from djem.utils.mon import M, Mon, mon  # NOQA
 from djem.utils.inspect import pp  # NOQA